Closes bug #41
This is a weird one. With the existing code I could reproduce the bug but not debug it. After transpiling all JS code debugging started working and the bug was gone. Go figure...

At least I enabled strict null checks, which showed me a few weak points in the code.